Safari-611.1.10/LayoutTests/editing/selection/select-out-of-floated-non-editable-03.html

<!DOCTYPE html>
<html>
<body>
<div id="container"><div>block-level text before floats</div><div id="target" style="float: left; margin-right: 20px">first float*</div><div style="float: left; margin-right: 20px">second float*</div>inline-level text after the floats</div>
<p style="padding-top: 1em;">
To test manually, drag from the middle of the 'first float' div to the
right, until the space between the 'second float' and the inline-text.
The selection should include both floats enterely with a portion of the
inline-level box's background and not jump to the beginning.
</p>
<pre id="console"></pre>
<script src="../../resources/dump-as-markup.js"></script>
<script src="resources/select-out-of-floated-non-editable.js"></script>
<script>
runTest("float*second float*", 100);
</script>
</body>
</html>
