Safari-611.1.10/LayoutTests/editing/selection/select-out-of-floated-non-editable-03-expected.txt

block-level text before floats
first float*second float*inline-level text after the floats
To test manually, drag from the middle of the 'first float' div to the right, until the space between the 'second float' and the inline-text. The selection should include both floats enterely with a portion of the inline-level box's background and not jump to the beginning.

SUCCESS: Selected test is 'float*second float*' as expected
| <div>
|   "block-level text before floats"
| <div>
|   id="target"
|   style="float: left; margin-right: 20px"
|   "first <#selection-anchor>float*"
| <div>
|   style="float: left; margin-right: 20px"
|   "second float*"
| "<#selection-focus>inline-level text after the floats"
